+++ server/src/main/java/com/oss/carbonadministrator/config/jwt/JwtAuthorizationFilter.java
@@ -3,8 +3,8 @@ package com.oss.carbonadministrator.config.jwt;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.oss.carbonadministrator.config.security.CustomUserDetails;
-import com.oss.carbonadministrator.domain.User;
-import com.oss.carbonadministrator.repository.UserRepository;
+import com.oss.carbonadministrator.domain.user.User;
+import com.oss.carbonadministrator.repository.user.UserRepository;
 import java.io.IOException;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
