+++ server/src/main/java/com/oss/carbonadministrator/controller/graph/GraphController.java
@@ -1,4 +1,4 @@
-package com.oss.carbonadministrator.controller;
+package com.oss.carbonadministrator.controller.graph;
 
 import com.oss.carbonadministrator.dto.response.ResponseDto;
 import com.oss.carbonadministrator.service.graph.GraphService;
