+++ server/src/main/java/com/oss/carbonadministrator/controller/image/ImageController.java
@@ -1,10 +1,10 @@
-package com.oss.carbonadministrator.controller;
+package com.oss.carbonadministrator.controller.image;
 
-import com.oss.carbonadministrator.domain.Electricity;
+import com.oss.carbonadministrator.domain.electricity.Electricity;
 import com.oss.carbonadministrator.dto.request.image.ElecImgRequest;
 import com.oss.carbonadministrator.dto.request.image.ImageRequest;
 import com.oss.carbonadministrator.dto.response.ResponseDto;
-import com.oss.carbonadministrator.service.image.ImageService;
+import com.oss.carbonadministrator.service.image.ElecImageService;
 import java.io.IOException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +23,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ImageController {
 
-    private final ImageService imageService;
+    private final ElecImageService elecImageService;
 
     @PostMapping("/electricity")
     @ResponseStatus(HttpStatus.CREATED)
@@ -32,12 +32,12 @@ public class ImageController {
 
         UUID uuid = UUID.randomUUID();
 
-        imageService.makeBase64ToImage(request.getImage(), ".jpg", uuid);
+        elecImageService.makeBase64ToImage(request.getImage(), ".jpg", uuid);
 
-        imageService.imageToJson(uuid.toString());
-        Electricity recognizedElecData = imageService.jsonToDto(uuid.toString());
+        elecImageService.imageToJson(uuid.toString());
+        Electricity recognizedElecData = elecImageService.jsonToDto(uuid.toString());
         // 데이터 저장 후 json 파일 삭제
-        Electricity savedData = imageService.save(request.getEmail(), request.getYear(), request.getMonth(), recognizedElecData);
+        Electricity savedData = elecImageService.save(request.getEmail(), request.getYear(), request.getMonth(), recognizedElecData);
 
         return ResponseDto.success(savedData, "전기 고지서 데이터 인식 및 저장 성공");
     }
@@ -45,13 +45,13 @@ public class ImageController {
     @PutMapping("/electricity/{electricityId}/edit")
     public ResponseDto editElecImgData(@PathVariable("electricityId") Long electricityId,
         @RequestBody ElecImgRequest requestDto) {
-        imageService.update(electricityId, requestDto.toElecEntity(requestDto));
+        elecImageService.update(electricityId, requestDto.toElecEntity(requestDto));
         return ResponseDto.success(null, "전기 고지서 데이터 사용자 수정 완료");
     }
 
     @PostMapping("/electricity/input")
     public ResponseDto inputElecData(@RequestBody ElecImgRequest requestDto) {
-        imageService.save(requestDto.getEmail(), requestDto.getYear(), requestDto.getMonth(),
+        elecImageService.save(requestDto.getEmail(), requestDto.getYear(), requestDto.getMonth(),
             requestDto.toElecEntity(requestDto));
         return ResponseDto.success(null, "사용자 데이터 직접 입력 성공");
     }
