+++ server/src/main/java/com/oss/carbonadministrator/controller/user/UserController.java
@@ -1,12 +1,12 @@
-package com.oss.carbonadministrator.controller;
+package com.oss.carbonadministrator.controller.user;
 
 import com.oss.carbonadministrator.dto.request.user.SignUpRequest;
 import com.oss.carbonadministrator.dto.request.user.UserEmailRequest;
 import com.oss.carbonadministrator.dto.request.user.UserNicknameRequest;
 import com.oss.carbonadministrator.dto.response.ResponseDto;
 import com.oss.carbonadministrator.dto.response.user.SignupResponse;
-import com.oss.carbonadministrator.exception.AlreadyExistEmailException;
-import com.oss.carbonadministrator.exception.AlreadyExistNicknameException;
+import com.oss.carbonadministrator.exception.user.AlreadyExistEmailException;
+import com.oss.carbonadministrator.exception.user.AlreadyExistNicknameException;
 import com.oss.carbonadministrator.service.user.UserService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
