+++ server/src/main/java/com/oss/carbonadministrator/domain/user/User.java
@@ -1,17 +1,13 @@
-package com.oss.carbonadministrator.domain;
+package com.oss.carbonadministrator.domain.user;
 
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.CascadeType;
+import com.oss.carbonadministrator.domain.base.BaseTimeEntity;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
