+++ server/src/main/java/com/oss/carbonadministrator/exception/image/ImgUploadFailException.java
@@ -1,4 +1,4 @@
-package com.oss.carbonadministrator.exception;
+package com.oss.carbonadministrator.exception.image;
 
 public class ImgUploadFailException extends RuntimeException {
