+++ server/src/main/java/com/oss/carbonadministrator/exception/user/AlreadyExistEmailException.java
@@ -1,4 +1,4 @@
-package com.oss.carbonadministrator.exception;
+package com.oss.carbonadministrator.exception.user;
 
 public class AlreadyExistEmailException extends RuntimeException {
