+++ server/src/main/java/com/oss/carbonadministrator/repository/electricity/ElectricityRepository.java
@@ -1,6 +1,6 @@
-package com.oss.carbonadministrator.repository;
+package com.oss.carbonadministrator.repository.electricity;
 
-import com.oss.carbonadministrator.domain.Electricity;
+import com.oss.carbonadministrator.domain.electricity.Electricity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
