+++ server/src/main/java/com/oss/carbonadministrator/repository/user/UserRepository.java
@@ -1,6 +1,6 @@
-package com.oss.carbonadministrator.repository;
+package com.oss.carbonadministrator.repository.user;
 
-import com.oss.carbonadministrator.domain.User;
+import com.oss.carbonadministrator.domain.user.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
